'add-global-settings'
