+++ .gitignore
@@ -5,4 +5,4 @@
 !.gitkeep
 
 # city combinations
-/city_coms_segmentation/city_coms/*
+/path_crawler/data/*
